refatorando as calculadoras para BigDecimal e mudando o BD

+++ src/main/java/com/rh/folhaPagamento/dto/FuncionarioRequestDTO.java
@@ -3,6 +3,8 @@ package com.rh.folhaPagamento.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 
 @Getter
 @Setter
@@ -11,13 +13,13 @@ public class FuncionarioRequestDTO {
     private String cpf;
     private String cargo;
     private int dependentes;
-    private double salarioBase;
+    private BigDecimal salarioBase;
     private boolean aptoPericulosidade;
     private int grauInsalubridade;
     private boolean valeTransporte;
     private boolean valeAlimentacao;
-    private double valorVT;
-    private double valorVA;
+    private BigDecimal valorVT;
+    private BigDecimal valorVA;
 
 
     private String login;
